+++ application/core/MY_Model.php
@@ -5,7 +5,7 @@ require_once 'DataMapper.php';
 
 /**
  * Generic data access model, for an RDB.
- * 
+ *
  * This class is called MY_Model to keep CodeIgniter happy.
  *
  * @author		JLP
@@ -255,7 +255,7 @@ class MY_Model2 extends MY_Model
 //  Record-oriented functions
 //---------------------------------------------------------------------------
 	// Retrieve an existing DB record as an object
-	function get($key1, $key2)
+	function get($key1, $key2 = null)
 	{
 		$this->db->where($this->_keyField, $key1);
 		$this->db->where($this->_keyField2, $key2);
@@ -286,7 +286,7 @@ class MY_Model2 extends MY_Model
 	}
 
 	// Delete a record from the DB
-	function delete($key1, $key2)
+	function delete($key1, $key2 = null)
 	{
 		$this->db->where($this->_keyField, $key1);
 		$this->db->where($this->_keyField2, $key2);
@@ -294,7 +294,7 @@ class MY_Model2 extends MY_Model
 	}
 
 	// Determine if a key exists
-	function exists($key1, $key2)
+	function exists($key1, $key2 = null)
 	{
 		$this->db->where($this->_keyField, $key1);
 		$this->db->where($this->_keyField2, $key2);