+++ application/models/Categories.php
@@ -1,32 +1,37 @@
 <?php
 
-/**
- * Generic data access abstraction.
- *
- * This is the intended implmentation for collection access.
- * Implementing classes will handle persistence.
- *
- * @author		JLP
- * @copyright           Copyright (c) 2010-2017, James L. Parry
- * ------------------------------------------------------------------------
- */
-interface DataMapper
+class Categories extends CSV_Model
 {
-//---------------------------------------------------------------------------
-//  Utility methods
-//---------------------------------------------------------------------------
+			//pk
+	    public $id;
+
+	    public $name;
+
+	    public $description;
+
+	public function __construct()
+	{
+		parent::__construct();
+
+	}
 
 	/**
 	 * Return the number of records in this table.
 	 * @return int The number of records in this table
 	 */
-	function size();
+	function size()
+	{
+
+	}
 
 	/**
 	 * Return the field names in this table, from the table metadata.
 	 * @return array(string) The field names in this table
 	 */
-	function fields();
+	function fields()
+	{
+
+	}
 
 //---------------------------------------------------------------------------
 //  C R U D methods
@@ -36,14 +41,20 @@ interface DataMapper
 	 * Only use this method if intending to create an empty record and then populate it.
 	 * @return object   An object with all its fields in place.
 	 */
-	function create();
+	function create()
+	{
+
+	}
 
 	/**
 	 * Add a record to the DB.
 	 * Request fails if the record already exists.
 	 * @param mixed $record The record to add, either an object or an associative array.
 	 */
-	function add($record);
+	function add($record)
+	{
+
+	}
 
 	/**
 	 * Retrieve an existing DB record as an object.
@@ -51,14 +62,20 @@ interface DataMapper
 	 * @param string $key2 Second part of composite key, if applicable
 	 * @return object The requested record, null if not found.
 	 */
-	function get($key, $key2 = null);
+	function get($key, $key2 = null)
+	{
+
+	}
 
 	/**
 	 * Update an existing DB record.
 	 * Method fails if the record does not exist.
 	 * @param mixed $record The record to update, either an object or an associative array.
 	 */
-	function update($record);
+	function update($record)
+	{
+
+	}
 
 	/**
 	 * Delete an existing DB record.
@@ -66,7 +83,10 @@ interface DataMapper
 	 * @param string $key Primary key of the record to delete.
 	 * @param string $key2 Second part of composite key, if applicable
 	 */
-	function delete($key, $key2 = null);
+	function delete($key, $key2 = null)
+	{
+
+	}
 
 	/**
 	 * Determine if a record exists.
@@ -74,58 +94,27 @@ interface DataMapper
 	 * @param string $key2 Second part of composite key, if applicable
 	 * @return boolean True if the record exists, false otherwise.
 	 */
-	function exists($key, $key2 = null);
+	function exists($key, $key2 = null)
+	{
+
+	}
 
 	/**
 	 * Determine the highest key used.
 	 * @return string The highest key
 	 */
-	function highest();
+	function highest()
+	{
+
+	}
 
 	/**
 	 * Return the first record
 	 * @return object The first record
 	 */
-	function first();
+	function first()
+	{
 
-//---------------------------------------------------------------------------
-//  Aggregate methods
-//---------------------------------------------------------------------------
-	/**
-	 * Retrieve all DB records.
-	 * @return array(object) All the records in the table.
-	 */
-	function all();
+	}
 
-	/**
-	 * Retrieve all DB records, but as a result set.
-	 * @return mixed The DB query result set.
-	 */
-	function results();
-
-	/**
-	 * Retrieve some of the DB records, namely those for which the
-	 * value of the field $what matches $which.
-	 * @param string    $what   Name of the field being matched.
-	 * @param   mixed   $which  Value sought.
-	 * @return mixed The selected records, as an array of records
-	 */
-	function some($what, $which);
-
-	/**
-	 * Retrieve records from the beginning of a table.
-	 * @return array(object) The relevant records
-	 */
-	function head($count);
-
-	/**
-	 * Retrieve records from the end of a table.
-	 * @return array(object) The relevant records
-	 */
-	function tail($count);
-
-	/**
-	 * Clear out everything
-	 */
-	function truncate();
 }